Creates a workingdir directory & adds missing files
Moves the readme file from document to workingdir. Adds example.bat and
default images to workingdir.

It should now be easier to put together a release since you can just add
all the required .dll files from Qt and the .exe files to workingdir and
then it should be ready for release.